init splitspense project using material design lib
