+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
 npm-debug.log
 build
+.idea/
+.idea_modules/
