+++ README.md
@@ -1,6 +1,14 @@
-# react-toolbox-example
+# Splitspense
+
+### About
+- Split bills among friends easily while you are on a trip
+- Keep track amount you and your friends spend and owned hustle free
+
+### Goal
+- Project to practice React and Firebase while having fun. :)
 
 ### Getting Started
 1. Clone this repository
 2. Run `npm install && npm start`
-3. Visit `0.0.0.0:8080` in your browser
+3. Visit `localhost:8080` in your browsers
+
