+++ src/app/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 // import 'react-toolbox/lib/commons.scss';           // Import common styles
-import PurpleAppBar from './PurpleAppBar.js';      // AppBar with simple overrides
-import SuccessButton from './SuccessButton.js';    // A button with complex overrides
+import PurpleAppBar from './PurpleAppBar.jsx';      // AppBar with simple overrides
+import SuccessButton from './SuccessButton.jsx';    // A button with complex overrides
 import { Button } from 'react-toolbox/lib/button'; // Bundled component import
 
 const App = () => (
