+++ src/app/client.jsx
@@ -1,6 +1,6 @@
 import 'react-toolbox/lib/commons.scss';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App.js';
+import App from './App.jsx';
 
 ReactDOM.render(<App />, document.getElementById('app'));
